requirements.txt
fastapi
uvicorn
streamlit
text-generation
transformers
python-dotenv
pydantic
requests
